Stop running the DHT after we have enough peers
Stop running the DHT if the peer count reaches two times the limit. This saves network bandwidth at the beginning of the stream, while still allowing for enough peers for swapping.

Not completely sure about the benefits, should be quite beneficial in theory, and it feels like it in practice.