More aggressive reconnect_wait
A reconnect wait like this is very negative behaviour - it leads to peers which are un-available being tried over and over again, taking up slots for new connections.

It's pretty safe to assume that if a peer fails the re-connect three times, it's not going to succeed in general. Better try new connections than keep re-trying.